refactor(ascend): override rms_norm_w8a16_fp8 in the backend
Move the Ascend W8A16 INT8 kernel behind the existing NVIDIA symbol so
SpecOpRegistrar covers it, and drop the extra ops/test entry points.
