+++ src/flag_gems/ops/__init__.py
@@ -730,10 +730,6 @@ from flag_gems.ops.resolve_conj import resolve_conj
 from flag_gems.ops.resolve_neg import resolve_neg
 from flag_gems.ops.rms_norm import rms_norm, rms_norm_backward, rms_norm_forward
 from flag_gems.ops.rms_norm_w8a16_fp8 import rms_norm_w8a16_fp8
-from flag_gems.ops.rmsnorm_w8a16_ascend import (
-    rms_norm_fp8_w8a16,
-    rms_norm_w8a16_ascend,
-)
 from flag_gems.ops.rnn_relu import rnn_relu
 from flag_gems.ops.roll import roll
 from flag_gems.ops.rot90 import rot90
@@ -1754,8 +1750,6 @@ __all__ = [
     "rms_norm_backward",
     "rms_norm_forward",
     "rms_norm_w8a16_fp8",
-    "rms_norm_fp8_w8a16",
-    "rms_norm_w8a16_ascend",
     "te_rmsnorm_bwd",
     "te_rmsnorm_fwd",
     "rnn_relu",
