+++ src/flag_gems/runtime/backend/_ascend/ops/__init__.py
@@ -100,12 +100,7 @@ from .replication_pad2d_backward import (
 )
 from .resolve_neg import resolve_neg
 from .rms_norm import rms_norm
-from flag_gems.ops.rmsnorm_w8a16_ascend import (
-    rms_norm_fp8_w8a16,
-    rms_norm_w8a16_ascend,
-)
-
-rms_norm_w8a16_fp8 = rms_norm_w8a16_ascend
+from .rms_norm_w8a16_fp8 import rms_norm_w8a16_fp8
 from .scatter import scatter, scatter_
 from .scatter_add_ import scatter_add_
 from .select_backward import select_backward
@@ -245,8 +240,6 @@ __all__ = [
     "replication_pad2d_backward_grad_input",
     "resolve_neg",
     "rms_norm",
-    "rms_norm_fp8_w8a16",
-    "rms_norm_w8a16_ascend",
     "rms_norm_w8a16_fp8",
     "scatter",
     "scatter_",
